Git: Address a technical review concern partially
Feedback from technical reviewer (KM) was:
"define DAG" directly above "Git stores the content as a directed acyclic.."
I partially addressed this by elaborating more on the practical difference
between a delta based changeset and a DAG content representation. Since this
book is being pitched to more experienced software developers I didn't want
to have a formal definition of a DAG and it to read like an undergraduate
text book. But I am open to suggestsions.

git-svn-id: http://www.aosabook.org/svn/volume2@880 dd5aa6cc-de63-413d-92ad-03a14aac4e6a